lift is_below_pair_threshold to NC level class, thus duplicate for heavy and asy
